+++ src/test/java/kr/hhplus/be/server/integration/domain/order/service/OrderServiceTest.java
@@ -2,6 +2,7 @@ package kr.hhplus.be.server.integration.domain.order.service;
 
 import kr.hhplus.be.server.domain.order.domain.Order;
 import kr.hhplus.be.server.domain.order.domain.OrderStatus;
+import kr.hhplus.be.server.domain.order.exception.OrderNotFoundException;
 import kr.hhplus.be.server.domain.order.service.OrderService;
 import kr.hhplus.be.server.domain.order.service.dto.OrderDetailDto;
 import kr.hhplus.be.server.domain.product.domain.Category;
@@ -23,6 +24,8 @@ import org.springframework.test.context.ActiveProfiles;
 import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
+import static org.assertj.core.api.AssertionsForClassTypes.assertThatCode;
 
 @ActiveProfiles("test")
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.NONE)
@@ -86,5 +89,46 @@ class OrderServiceTest {
         // then
         assertThat(foundOrder.getUser().getId()).isEqualTo(savedOrder.getUser().getId());
         assertThat(foundOrder.getStatus()).isEqualTo(OrderStatus.PAYMENT_PENDING);
+    }
+
+    @DisplayName("Order 소유자 검증 검증 - 성공")
+    @Test
+    void validateOrderOwnership() {
+        // given
+        User user1 = UserFixture.USER(1L);
+
+        String name = "라넌큘러스 오버핏 맨투맨";
+        Category category = CategoryFixture.create("상의");
+        int price = 12_000;
+        int stockQuantity = 100;
+        Product product = new Product(name, category, price, stockQuantity);
+
+        Order order = new Order(user1);
+        order.addOrderProduct(product, 1);
+
+        // when & then
+        assertThatCode(() -> orderService.validateOrderOwnership(user1, order))
+                .doesNotThrowAnyException();
+    }
+
+    @DisplayName("Order 소유자 검증 검증 - 실패 - 유저가 생성한 주문이 아닌 경우 예외 발생")
+    @Test
+    void validateOrderOwnership_Fail_NotMine() {
+        // given
+        User user1 = UserFixture.USER(1L);
+        User user2 = UserFixture.USER(2L);
+
+        String name = "라넌큘러스 오버핏 맨투맨";
+        Category category = CategoryFixture.create("상의");
+        int price = 12_000;
+        int stockQuantity = 100;
+        Product product = new Product(name, category, price, stockQuantity);
+
+        Order order = new Order(user1);
+        order.addOrderProduct(product, 1);
+
+        // when & then
+        assertThatThrownBy(() -> orderService.validateOrderOwnership(user2, order))
+                .isInstanceOf(OrderNotFoundException.class);
     }
 }
