+++ src/test/java/kr/hhplus/be/server/unit/domain/order/service/OrderServiceTest.java
@@ -1,6 +1,7 @@
 package kr.hhplus.be.server.unit.domain.order.service;
 
 import kr.hhplus.be.server.domain.order.domain.Order;
+import kr.hhplus.be.server.domain.order.exception.OrderNotFoundException;
 import kr.hhplus.be.server.domain.order.repository.OrderCoreRepository;
 import kr.hhplus.be.server.domain.order.service.OrderService;
 import kr.hhplus.be.server.domain.order.service.dto.OrderDetailDto;
@@ -19,6 +20,8 @@ import org.mockito.junit.jupiter.MockitoExtension;
 import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
+import static org.assertj.core.api.AssertionsForClassTypes.assertThatCode;
 import static org.mockito.Mockito.*;
 
 @ExtendWith(MockitoExtension.class)
@@ -80,5 +83,46 @@ class OrderServiceTest {
         assertThat(savedOrder.getOrderProducts().size()).isZero();
 
         verify(orderCoreRepository, times(1)).findById(1L);
+    }
+
+    @DisplayName("Order 소유자 검증 - 성공")
+    @Test
+    void validateOrderOwnership() {
+        // given
+        User user1 = UserFixture.USER(1L);
+
+        String name = "라넌큘러스 오버핏 맨투맨";
+        Category category = CategoryFixture.create("상의");
+        int price = 12_000;
+        int stockQuantity = 100;
+        Product product = new Product(name, category, price, stockQuantity);
+
+        Order order = new Order(user1);
+        order.addOrderProduct(product, 1);
+
+        // when & then
+        assertThatCode(() -> orderService.validateOrderOwnership(user1, order))
+                .doesNotThrowAnyException();
+    }
+
+    @DisplayName("Order 소유자 검증 - 실패 - 유저가 생성한 주문이 아닌 경우 예외 발생")
+    @Test
+    void validateOrderOwnership_Fail_NotMine() {
+        // given
+        User user1 = UserFixture.USER(1L);
+        User user2 = UserFixture.USER(2L);
+
+        String name = "라넌큘러스 오버핏 맨투맨";
+        Category category = CategoryFixture.create("상의");
+        int price = 12_000;
+        int stockQuantity = 100;
+        Product product = new Product(name, category, price, stockQuantity);
+
+        Order order = new Order(user1);
+        order.addOrderProduct(product, 1);
+
+        // when & then
+        assertThatThrownBy(() -> orderService.validateOrderOwnership(user2, order))
+                .isInstanceOf(OrderNotFoundException.class);
     }
 }
